Add files via upload
These two notebooks were used for a general chemistry lab course. It was taught online during Fall 2020. The 2 different notebooks are meant to be assigned to every other student (half the class gets ozone, half the class gets NO2) so that the students who have different notebooks will be paired up to discuss similarities and differences in their outcomes/code.